PSP Integration candidate: Equuleus-rc1+dev18

+++ fsw/modules/eeprom_mmap_file/cfe_psp_eeprom_mmap_file.c
@@ -190,7 +190,7 @@ void eeprom_mmap_file_Init(uint32 PspModuleId)
         Status = CFE_PSP_MemRangeSet(1, CFE_PSP_MEM_EEPROM, eeprom_address, eeprom_size, CFE_PSP_MEM_SIZE_DWORD,
                                      CFE_PSP_MEM_ATTR_READWRITE);
         OS_printf("CFE_PSP: EEPROM Range (2) created: Start Address = %08lX, Size = %08X Status = %d\n",
-                  (unsigned long)eeprom_address, (unsigned int)eeprom_size, Status);
+                  (unsigned long)eeprom_address, (unsigned int)eeprom_size, (int)Status);
     }
     else
     {
